Added in player score and made the regular jump smaller
