+++ Assets/Scripts/Player/PlayerMovement.cs
@@ -11,7 +11,6 @@ public class PlayerMovement : MonoBehaviour
     [SerializeField] private float _moveSpeed = 10.0f;
     [SerializeField] private float _jumpPower = 25.0f;
     [SerializeField] private float _maxDistance = 1.0f;
-    [SerializeField] private float _jumpTime = 0.35f;
 
     [Header("Ground Check")]
     //The Layermask is what the ground check will be checking for
@@ -63,7 +62,7 @@ public class PlayerMovement : MonoBehaviour
         // Whenever button is released, cut the y velocity in half.
         if (context.canceled && _playerRB.velocity.y > 0f)
         {
-            _playerRB.velocity = new Vector2(_playerRB.velocity.x, _playerRB.velocity.y * 0.75f);
+            _playerRB.velocity = new Vector2(_playerRB.velocity.x, _playerRB.velocity.y * 0.3f);
         }
     }
