--- Assets/Scripts/Player/PlayerScore.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b4c8382305e18614cb78065296e03ca5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
